# Docs/Api/cpay-v2-openapi.yaml
openapi: 3.0.3
info:
  title: CPay Gateway API
  version: 1.6.0
  description: Merchant payment, callback operations, reconciliation, finance, and operations endpoints for CPay.
security:
  - CPayV2Signature: []
paths:
...
  /api/v2/admin/recon-finance/summary:
    get:
      summary: Get reconciliation finance summary
      operationId: reconciliationFinanceSummary
      security:
        - AdminBasicAuth: []
      responses:
        '200':
          description: Finance summary
  /api/v2/admin/recon-finance/close:
    post:
      summary: Submit a reconciliation daily close for maker-checker approval (maker step)
      operationId: reconciliationDailyClose
      security:
        - AdminBasicAuth: []
      parameters:
        - name: date
          in: query
          required: true
          schema:
            type: string
            format: date
        - name: currency
          in: query
          required: false
          schema:
            type: string
            default: UGX
        - name: submittedBy
          in: query
          required: false
          schema:
            type: string
            default: system
      responses:
        '200':
          description: Daily close row id in PENDING_APPROVAL
  /api/v2/admin/recon-finance/close/approve:
    post:
      summary: Checker approval for a submitted daily close
      operationId: approveReconciliationDailyClose
      security:
        - AdminBasicAuth: []
      parameters:
        - name: date
          in: query
          required: true
          schema:
            type: string
            format: date
        - name: currency
          in: query
          required: false
          schema:
            type: string
            default: UGX
        - name: approvedBy
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Daily close approved and closed
        '400':
          $ref: '#/components/responses/BadRequest'
  /api/v2/admin/recon-finance/close/reject:
    post:
      summary: Checker rejection for a submitted daily close
      operationId: rejectReconciliationDailyClose
      security:
        - AdminBasicAuth: []
      parameters:
        - name: date
          in: query
          required: true
          schema:
            type: string
            format: date
        - name: currency
          in: query
          required: false
          schema:
            type: string
            default: UGX
        - name: rejectedBy
          in: query
          required: true
          schema:
            type: string
        - name: reason
          in: query
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Daily close marked rejected and left pending for re-submit
        '400':
          $ref: '#/components/responses/BadRequest'
...
  /api/v2/admin/payout-approvals:
    get:
      summary: List payouts awaiting maker-checker approval
      operationId: listPendingPayoutApprovals
      security:
        - AdminBasicAuth: []
      parameters:
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 100
      responses:
        '200':
          description: Pending payout approvals
  /api/v2/admin/payout-approvals/{queueId}/approve:
    post:
      summary: Checker approval for a queued payout; re-executes the payout through the normal path
      operationId: approveQueuedPayout
      security:
        - AdminBasicAuth: []
      parameters:
        - name: queueId
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: approvedBy
          in: query
          required: true
          schema:
            type: string
      responses:
        '202':
          description: Payout approved and submitted through the existing payment executor
        '400':
          $ref: '#/components/responses/BadRequest'
  /api/v2/admin/payout-approvals/{queueId}/reject:
    post:
      summary: Checker rejection for a queued payout
      operationId: rejectQueuedPayout
      security:
        - AdminBasicAuth: []
      parameters:
        - name: queueId
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: rejectedBy
          in: query
          required: true
          schema:
            type: string
        - name: reason
          in: query
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Payout rejected
        '400':
          $ref: '#/components/responses/BadRequest'
  /api/v2/admin/payout-approvals/{queueId}/cancel:
    post:
      summary: Cancel a queued payout before decision
      operationId: cancelQueuedPayout
      security:
        - AdminBasicAuth: []
      parameters:
        - name: queueId
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: cancelledBy
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Payout cancelled
        '400':
          $ref: '#/components/responses/BadRequest'
...
components:
  securitySchemes:
    CPayV2Signature:
      type: apiKey
      in: header
      name: X-CPay-Signature
    AdminBasicAuth:
      type: http
      scheme: basic
...
  responses:
    BadRequest:
      description: Request was rejected
...